Amend the Loki article to include parser expressions
The "Displaying geohash tags from a Loki datasource in a Grafana
Worldmap Panel" article is slightly outdated considering that previously
the only way of using the geohash information was through a label, which
is hihgly inefficient. The new parser expressions allow to use the
information from the actual payload without polluting the label set.